Finished refactoring to use int64 as base currency, however market position calculations not working, slice not lined up
