+++ backend/main.go
@@ -5,6 +5,7 @@ import (
 	"net/http"
 	"socialpredict/middleware"
 	"socialpredict/migration"
+	"socialpredict/seed"
 	"socialpredict/server"
 	"socialpredict/util"
 )
@@ -30,9 +31,9 @@ func main() {
 	migration.MigrateDB(db)
 
 	// Seed the admin user
-	// seed.SeedUsers(db)
-	// seed.SeedMarket(db)
-	// seed.SeedBets(db)
+	seed.SeedUsers(db)
+	seed.SeedMarket(db)
+	seed.SeedBets(db)
 
 	server.Start()
 }
